
\name{UNDOCUMENTED}
\alias{UNDOCUMENTED}
\title{Undocumented or minimally documented functions }
\description{\itemize{
\item adm.std 
\item admnormapprox 
\item admprofile 
\item brewer.all.info 
\item bumppointers 
\item cmd 
\item examine.brewer 
\item fix.oracle.input <- function(list) -- fixes stuff from oracle 
\item getcol 
\item jpegfile <- function (fname = "R-plot", quality = 100, epsdir = TRUE) 
\item legend.pk 
\item legend2 
\item lsf 
\item lsv 
\item missingfiles 
\item mybarplot 
\item myfunction 
\item paletteTest 
\item pielet <- function(x,y,r,a,col,...) -- draws a pie chart at x,y
\item plot0 <- function(x, y, ylim = range(0, y), ...) ie. quick plot with 0 in y-range
\item plotSymbols <- function() -- plots the various choices of plotting symbols
\item showsymbols <- function() -- plots the various choices of plotting symbols 
\item rwiki.search 
\item sumna <- function(x){sum(is.na(x))}
\item table3 
\item \%&\% 
\item \%*=\% 
\item \%+=\% 
\item \%-=\% 
\item \%/=\% 
}}
