\name{sizelist}
\alias{sizelist}
\title{ Object sizes }
\description{
  Produces sorted list of objects with approx. memory sizes.
}
\usage{
sizelist(env=1)
}
\arguments{
  \item{env}{ -- environment (integer or name) }
}
\value{
  Single column data frame containing the object list
}
\author{ P. Kleiber }
